+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added cmdlet `Convert-ObjectToHashtable` wich can be used to convert any object
+ to hashtable. It accepts objects from pipeline.
+
 ## [0.10.2] - 2021-03-24
 
 ### Changed